app/pyproject.toml
[tool.poetry]
name = "assetledger"
version = "0.1.0"
description = "Asset Ledger - Bronze/Silver/Gold pipeline for IT asset tracking"
authors = ["Asset Team"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
psycopg2-binary = "^2.9"
gunicorn = "^21.2"
google-api-python-client = "^2.100"
google-auth = "^2.23"
google-auth-httplib2 = "^0.1"
google-auth-oauthlib = "^1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
pytest-django = "^4.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
